Improve boid flocker model
