+++ examples/boid_flockers/Readme.md
@@ -1,34 +1,47 @@
-# Flockers
+# Boids Flockers
+
+## Summary
 
 An implementation of Craig Reynolds's Boids flocker model. Agents (simulated birds) try to fly towards the average position of their neighbors and in the same direction as them, while maintaining a minimum distance. This produces flocking behavior.
 
 This model tests Mesa's continuous space feature, and uses numpy arrays to represent vectors. It also demonstrates how to create custom visualization components.
 
+## Installation
+
+To install the dependencies use pip and the requirements.txt in this directory. e.g.
+
+```
+    $ pip install -r requirements.txt
+```
+
 ## How to Run
 
-Launch the model:
-```
-    $ python Flocker_Server.py
-```
+* To launch the visualization interactively, run ``mesa runserver`` in this directory. e.g.
 
-Then open your browser to [http://127.0.0.1:8521/](http://127.0.0.1:8521/) and press Reset, then Run.
-
-## Files
-
-* [flockers/model.py](flockers/model.py): Core model file; contains the BoidModel class.
-* [flockers/boid.py](flockers/boid.py): The Boid agent class.
-* [flockers/SimpleContinuousModule.py](flockers/SimpleContinuousModule.py): Defines ``SimpleCanvas``, the Python side of a custom visualization module for drawing agents with continuous positions.
-* [flockers/simple_continuous_canvas.js](flockers/simple_continuous_canvas.js): JavaScript side of the ``SimpleCanvas`` visualization module; takes the output generated by the Python ``SimpleCanvas`` element and draws it in the browser window via HTML5 canvas.
-* [flockers/server.py](flockers/server.py): Sets up the visualization; uses the SimpleCanvas element defined above
-* [run.py](run.py) Launches the visualization.
-* [Flocker Test.ipynb](Flocker Test.ipynb): Tests the model in a Jupyter notebook.
-
-## Further Reading
-
-=======
-* Launch the visualization
 ```
 $ mesa runserver
 ```
-* Visit your browser: http://127.0.0.1:8521/
-* In your browser hit *run*
+
+or 
+
+Directly run the file ``app.py`` in the terminal. e.g.
+
+```
+    $ python app.py
+```
+
+* Then open your browser to [http://127.0.0.1:8521/](http://127.0.0.1:8521/) and press Reset, then Run.
+
+## Files
+
+* [boid_flockers/model.py](boid_flockers/model.py): Core model file; contains the Boid Model and Boid Agent class.
+* [boid_flockers/SimpleContinuousModule.py](boid_flockers/SimpleContinuousModule.py): Defines ``SimpleCanvas``, the Python side of a custom visualization module for drawing agents with continuous positions.
+* [boid_flockers/simple_continuous_canvas.js](boid_flockers/simple_continuous_canvas.js): JavaScript side of the ``SimpleCanvas`` visualization module; takes the output generated by the Python ``SimpleCanvas`` element and draws it in the browser window via HTML5 canvas.
+* [boid_flockers/server.py](boid_flockers/server.py): Sets up the visualization; uses the SimpleCanvas element defined above
+* [run.py](run.py) Launches the visualization.
+* [Flocker_Test.ipynb](Flocker_Test.ipynb): Tests the model in a Jupyter notebook.
+
+## Further Reading
+
+The following link can be visited for more information on the boid flockers model:
+https://cs.stanford.edu/people/eroberts/courses/soco/projects/2008-09/modeling-natural-systems/boids.html
